Use topic instead of fanout exchange
The goal was to find a way to bind the temporary queue with multiple routing
keys.  Temporary queues are a `rabbitmq-native` plugin specific feature.  They
are not part of RabbitMQ itself.  The plugin is too simplisting in its notation
for bindings to enable multiple bindings.

I ended up putting the consumer logic in a base class under `src/` and having
simple shells under `rabbit-consumers/` that extend the base class and supply
a binding.  I have one shell per routing key.  I end up with multiple consumers,
but at least the code is not duplicated.

Incidentally, I also pushed the consumer logic to a service.  One more level of
indirection, FTW!